refactor: update the main.go script to use a cmder
