+++ hyperlane/cmd/hyp/cmd/events.go
@@ -1,0 +1,90 @@
+package cmd
+
+import (
+	"log"
+
+	"github.com/bcp-innovations/hyperlane-cosmos/util"
+	ismtypes "github.com/bcp-innovations/hyperlane-cosmos/x/core/01_interchain_security/types"
+	hooktypes "github.com/bcp-innovations/hyperlane-cosmos/x/core/02_post_dispatch/types"
+	coretypes "github.com/bcp-innovations/hyperlane-cosmos/x/core/types"
+	warptypes "github.com/bcp-innovations/hyperlane-cosmos/x/warp/types"
+	abci "github.com/cometbft/cometbft/abci/types"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/gogoproto/proto"
+)
+
+func parseISMFromEvents(events []abci.Event) util.HexAddress {
+	var ismID util.HexAddress
+	for _, evt := range events {
+		if evt.GetType() == proto.MessageName(&ismtypes.EventCreateNoopIsm{}) {
+			event, err := sdk.ParseTypedEvent(evt)
+			if err != nil {
+				log.Fatalf("failed to parse typed event: %v", err)
+			}
+
+			if ismEvent, ok := event.(*ismtypes.EventCreateNoopIsm); ok {
+				log.Printf("successfully created Noop ISM: %s\n", ismEvent)
+				ismID = ismEvent.IsmId
+			}
+		}
+	}
+
+	return ismID
+}
+
+func parseHooksIDFromEvents(events []abci.Event) util.HexAddress {
+	var hookID util.HexAddress
+	for _, evt := range events {
+		if evt.GetType() == proto.MessageName(&hooktypes.EventCreateNoopHook{}) {
+			event, err := sdk.ParseTypedEvent(evt)
+			if err != nil {
+				log.Fatalf("failed to parse typed event: %v", err)
+			}
+
+			if hookEvent, ok := event.(*hooktypes.EventCreateNoopHook); ok {
+				log.Printf("successfully created NoopHook: %s\n", hookEvent)
+				hookID = hookEvent.NoopHookId
+			}
+		}
+	}
+
+	return hookID
+}
+
+func parseMailboxIDFromEvents(events []abci.Event) util.HexAddress {
+	var mailboxID util.HexAddress
+	for _, evt := range events {
+		if evt.GetType() == proto.MessageName(&coretypes.EventCreateMailbox{}) {
+			event, err := sdk.ParseTypedEvent(evt)
+			if err != nil {
+				log.Fatalf("failed to parse typed event: %v", err)
+			}
+
+			if mailboxEvent, ok := event.(*coretypes.EventCreateMailbox); ok {
+				log.Printf("successfully created Mailbox: %s\n", mailboxEvent)
+				mailboxID = mailboxEvent.MailboxId
+			}
+		}
+	}
+
+	return mailboxID
+}
+
+func parseCollateralTokenIDFromEvents(events []abci.Event) util.HexAddress {
+	var tokenID util.HexAddress
+	for _, evt := range events {
+		if evt.GetType() == proto.MessageName(&warptypes.EventCreateCollateralToken{}) {
+			event, err := sdk.ParseTypedEvent(evt)
+			if err != nil {
+				log.Fatalf("failed to parse typed event: %v", err)
+			}
+
+			if tokenEvent, ok := event.(*warptypes.EventCreateCollateralToken); ok {
+				log.Printf("successfully created CollateralToken: %s\n", tokenEvent)
+				tokenID = tokenEvent.TokenId
+			}
+		}
+	}
+
+	return tokenID
+}
